requirements.txt
taxii2-client ==  2.2.2
stix2 == 2.0.2
flask == 1.1.2
requests == 2.24.0
pyyaml == 5.3.1
python-dateutil == 2.8.1
gevent==20.9.0